Split up parallel nodes tests.
